+++ src/main/java/com/example/library/persistence/repository/AuthorRepository.java
@@ -3,5 +3,5 @@ package com.example.library.persistence.repository;
 import com.example.library.persistence.entity.AuthorEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorRepository extends JpaRepository<AuthorEntity, Integer> {
+public interface AuthorRepository extends JpaRepository<AuthorEntity, Long> {
 }
