+++ src/main/java/com/example/library/persistence/repository/BookDetailRepository.java
@@ -3,5 +3,5 @@ package com.example.library.persistence.repository;
 import com.example.library.persistence.entity.BookDetailEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookDetailRepository extends JpaRepository<BookDetailEntity, Integer> {
+public interface BookDetailRepository extends JpaRepository<BookDetailEntity, Long> {
 }
