+++ src/main/java/com/example/library/persistence/repository/BookRepository.java
@@ -3,5 +3,5 @@ package com.example.library.persistence.repository;
 import com.example.library.persistence.entity.BookEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<BookEntity, Integer> {
+public interface BookRepository extends JpaRepository<BookEntity, Long> {
 }
