+++ src/main/java/com/example/library/persistence/repository/LoanRepository.java
@@ -3,5 +3,5 @@ package com.example.library.persistence.repository;
 import com.example.library.persistence.entity.LoanEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoanRepository extends JpaRepository<LoanEntity, Integer> {
+public interface LoanRepository extends JpaRepository<LoanEntity, Long> {
 }
