+++ src/main/java/com/example/library/persistence/repository/MemberRepository.java
@@ -3,5 +3,5 @@ package com.example.library.persistence.repository;
 import com.example.library.persistence.entity.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<MemberEntity, Integer> {
+public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
 }
